requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-core
tavily-python
python-dotenv
requests